Removed BaseCliFxException and refactored CliApplication accordingly.
